Fix extensions and retain field ordering
Extensions were broken in that inheritance was working in reverse,
i.e. base classes were inheriting from extending classes. Fix this
and also implement inheritance from SimpleType using the name "value"
for values. This is needed to meaningfully handle XML where not
everything is a named value of an element. Inheritance is done prior
to processing structures own fields.

Also retain ordering. This makes inheritance more transparent as
inherited types are included at the top of structures, recursively.
This also simplifies comparing the output to the XSD.